fix(telegram): show phased overlay + refresh bot info on Settings reconfigure
The Settings reconfigure flow showed only a spinner with 'Connecting...'
while the backend restarted the gateway and waited for it to come back —
which can take 10+ seconds. Users got no progress feedback and the bot
name/link in the configured card kept showing the old value after save.

Changes:
- Extract the phased progress overlay from TelegramStep into a shared
  TelegramConfiguringOverlay component (5 phases: token verified →
  connecting → restarting gateway → waiting for gateway health → ready).
- Use the shared overlay in both TelegramStep (setup wizard) and
  SettingsApp (desktop Settings app).
- Refactor the telegram/status fetch into a reusable refreshTelegramStatus
  callback and re-invoke it in the overlay's onDone, so the bot name,
  @username, and 'Open in Telegram' link update to reflect the newly
  configured bot without a page reload.
